Move get_account_alias to ef_conf_utils
