+++ efopen/ef_context.py
@@ -17,8 +17,9 @@ limitations under the License.
 """
 
 from ef_service_registry import EFServiceRegistry
-from ef_utils import get_account_alias, get_env_short, whereami
+from ef_utils import get_env_short, whereami
 from ef_conf_utils import env_valid
+from ef_conf_utils import get_account_alias
 from ef_conf_utils import global_env_valid
 
 class EFContext(object):
